Inline bottom navigation ad banner
